+++ NEWS.md
@@ -12,6 +12,10 @@ changed the format of the intervals specification.
 
 * Added Clast.pred related NCA calculations
 * Added N to summary of PKNCAresults
+* Added parameter selection between Clast,observed and Clast,predicted across all parameters
+* Enabled PKNCAdose to be specified with one-sided formula
+* Improved error reporting so that the group and time (interval specification) is reported in addition to the error.
+* PKNCAdose now allows route of administration and IV infusion parameters of rate/duration to be specified
 
 # PKNCA 0.7.1